+++ nix/materialized-x86_64-linux/.plan.nix/plutus-certification.nix
@@ -34,44 +34,48 @@
       "library" = {
         depends = [
           (hsPkgs."base" or (errorHandler.buildDepError "base"))
+          (hsPkgs."containers" or (errorHandler.buildDepError "containers"))
           (hsPkgs."servant" or (errorHandler.buildDepError "servant"))
+          (hsPkgs."servant-client" or (errorHandler.buildDepError "servant-client"))
+          (hsPkgs."servant-client-core" or (errorHandler.buildDepError "servant-client-core"))
+          (hsPkgs."servant-server" or (errorHandler.buildDepError "servant-server"))
+          (hsPkgs."conduit" or (errorHandler.buildDepError "conduit"))
+          (hsPkgs."cicero-api" or (errorHandler.buildDepError "cicero-api"))
           (hsPkgs."network-uri" or (errorHandler.buildDepError "network-uri"))
           (hsPkgs."bytestring" or (errorHandler.buildDepError "bytestring"))
           (hsPkgs."uuid" or (errorHandler.buildDepError "uuid"))
           (hsPkgs."text" or (errorHandler.buildDepError "text"))
+          (hsPkgs."mtl" or (errorHandler.buildDepError "mtl"))
+          (hsPkgs."http-client-tls" or (errorHandler.buildDepError "http-client-tls"))
           (hsPkgs."aeson" or (errorHandler.buildDepError "aeson"))
+          (hsPkgs."aeson-qq" or (errorHandler.buildDepError "aeson-qq"))
           ];
         buildable = true;
-        modules = [ "Plutus/Certification/API" ];
+        modules = [
+          "Paths_plutus_certification"
+          "Plutus/Certification/API"
+          "Plutus/Certification/Cache"
+          "Plutus/Certification/Cicero"
+          "Plutus/Certification/Client"
+          "Plutus/Certification/Server"
+          ];
         hsSourceDirs = [ "src" ];
         };
       exes = {
         "plutus-certification" = {
           depends = [
             (hsPkgs."base" or (errorHandler.buildDepError "base"))
-            (hsPkgs."uuid" or (errorHandler.buildDepError "uuid"))
-            (hsPkgs."containers" or (errorHandler.buildDepError "containers"))
             (hsPkgs."servant-server" or (errorHandler.buildDepError "servant-server"))
             (hsPkgs."optparse-applicative" or (errorHandler.buildDepError "optparse-applicative"))
             (hsPkgs."warp" or (errorHandler.buildDepError "warp"))
             (hsPkgs."transformers" or (errorHandler.buildDepError "transformers"))
-            (hsPkgs."cicero-api" or (errorHandler.buildDepError "cicero-api"))
-            (hsPkgs."servant-client" or (errorHandler.buildDepError "servant-client"))
             (hsPkgs."servant-client-core" or (errorHandler.buildDepError "servant-client-core"))
-            (hsPkgs."aeson-qq" or (errorHandler.buildDepError "aeson-qq"))
-            (hsPkgs."network-uri" or (errorHandler.buildDepError "network-uri"))
-            (hsPkgs."bytestring" or (errorHandler.buildDepError "bytestring"))
             (hsPkgs."streaming-commons" or (errorHandler.buildDepError "streaming-commons"))
             (hsPkgs."network" or (errorHandler.buildDepError "network"))
             (hsPkgs."async" or (errorHandler.buildDepError "async"))
-            (hsPkgs."http-client-tls" or (errorHandler.buildDepError "http-client-tls"))
-            (hsPkgs."mtl" or (errorHandler.buildDepError "mtl"))
-            (hsPkgs."conduit" or (errorHandler.buildDepError "conduit"))
-            (hsPkgs."aeson" or (errorHandler.buildDepError "aeson"))
             (hsPkgs."plutus-certification" or (errorHandler.buildDepError "plutus-certification"))
             ];
           buildable = true;
-          modules = [ "Paths_plutus_certification" ];
           hsSourceDirs = [ "server" ];
           mainPath = [ "Main.hs" ];
           };
